Iteration assignment
Found it incredibly difficult to work like this. Encountered multiple errors using the github thing (sorry to say so, but why not just keep doing the work like we did in assignment 1-4?). I think I might have gotten this correctly now, but it was incredibly frustrating. Some of the errors I encountered required multiple hours of googling to find solutions. Didn't work for me to set up the github desktop like explained in the videos.